+++ Models/LH/LHWWWVertex.cc
@@ -32,7 +32,7 @@ void LHWWWVertex::Init() {
 
 LHWWWVertex::LHWWWVertex() : _couplast(0.),_q2last(0.*GeV2) {
   // particles
-  vector<int> first,second,third;
+  vector<long> first,second,third;
   first.push_back(24);
   second.push_back(-24);
   third.push_back(22);
