+++ Models/StandardModel/SMWWWVertex.cc
@@ -71,7 +71,7 @@ void SMWWWVertex::setCoupling(Energy2 q2,tcPDPtr a,tcPDPtr b, tcPDPtr c) {
 
 SMWWWVertex::SMWWWVertex() : _zfact(0.),_couplast(0.), _q2last(0.*GeV2) {
   // particles
-  vector<int> first,second,third;
+  vector<long> first,second,third;
   first.push_back(24);
   second.push_back(-24);
   third.push_back(22);
